pyproject.toml
[tool.poetry]
name = "squared"
version = "0.1.0"
description = ""
authors = ["Allan Napier <49874689+allannapier@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
flask = "^3.1.0"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
